+++ src/openassetio-python/hostAPI/ManagerBinding.cpp
@@ -17,5 +17,11 @@ void registerManager(const py::module& mod) {
       .def("identifier", &Manager::identifier)
       .def("displayName", &Manager::displayName)
       .def("info", &Manager::info)
-      .def("initialize", &Manager::initialize);
+      .def("initialize", &Manager::initialize)
+      .def("createContext", &Manager::createContext)
+      .def("createChildContext", &Manager::createChildContext,
+           py::arg("parentContext").none(false))
+      .def("persistenceTokenForContext", &Manager::persistenceTokenForContext,
+           py::arg("context").none(false))
+      .def("contextFromPersistenceToken", &Manager::contextFromPersistenceToken, py::arg("token"));
 }