+++ src/openassetio-python/managerAPI/ManagerInterfaceBinding.cpp
@@ -3,7 +3,9 @@
 #include <pybind11/stl.h>
 
 #include <openassetio/InfoDictionary.hpp>
+#include <openassetio/managerAPI/HostSession.hpp>
 #include <openassetio/managerAPI/ManagerInterface.hpp>
+#include <openassetio/managerAPI/ManagerStateBase.hpp>
 #include <openassetio/typedefs.hpp>
 
 #include "../_openassetio.hpp"
@@ -31,8 +33,30 @@ struct PyManagerInterface : ManagerInterface {
     PYBIND11_OVERRIDE(InfoDictionary, ManagerInterface, info, /* no args */);
   }
 
-  void initialize(HostSessionPtr hostSession) override {
+  void initialize(const HostSessionPtr& hostSession) override {
     PYBIND11_OVERRIDE_PURE(void, ManagerInterface, initialize, hostSession);
+  }
+
+  ManagerStateBasePtr createState(const HostSessionPtr& hostSession) override {
+    PYBIND11_OVERRIDE(ManagerStateBasePtr, ManagerInterface, createState, hostSession);
+  }
+
+  ManagerStateBasePtr createChildState(const ManagerStateBasePtr& parentState,
+                                       const HostSessionPtr& hostSession) override {
+    PYBIND11_OVERRIDE(ManagerStateBasePtr, ManagerInterface, createChildState, parentState,
+                      hostSession);
+  }
+
+  std::string persistenceTokenForState(const ManagerStateBasePtr& parentState,
+                                       const HostSessionPtr& hostSession) override {
+    PYBIND11_OVERRIDE(std::string, ManagerInterface, persistenceTokenForState, parentState,
+                      hostSession);
+  }
+
+  ManagerStateBasePtr stateFromPersistenceToken(const std::string& token,
+                                                const HostSessionPtr& hostSession) override {
+    PYBIND11_OVERRIDE(ManagerStateBasePtr, ManagerInterface, stateFromPersistenceToken, token,
+                      hostSession);
   }
 };
 
@@ -50,5 +74,12 @@ void registerManagerInterface(const py::module& mod) {
       .def("identifier", &ManagerInterface::identifier)
       .def("displayName", &ManagerInterface::displayName)
       .def("info", &ManagerInterface::info)
-      .def("initialize", &ManagerInterface::initialize);
+      .def("initialize", &ManagerInterface::initialize)
+      .def("createState", &ManagerInterface::createState, py::arg("hostSession").none(false))
+      .def("createChildState", &ManagerInterface::createChildState,
+           py::arg("parentState").none(false), py::arg("hostSession").none(false))
+      .def("persistenceTokenForState", &ManagerInterface::persistenceTokenForState,
+           py::arg("state").none(false), py::arg("hostSession").none(false))
+      .def("stateFromPersistenceToken", &ManagerInterface::stateFromPersistenceToken,
+           py::arg("token"), py::arg("hostSession").none(false));
 }